:truck: move util tests to own module (mainly to mark module with `test_` so it's picked up by pytest !)
